Compiled all controls into one file and renamed file. Also reworked start screen
